+++ google/cloud/spanner_dbapi/connection.py
@@ -6,15 +6,13 @@
 
 """DB-API Connection for the Google Cloud Spanner."""
 
-# from google.cloud import spanner_v1
+from google.api_core.gapic_v1.client_info import ClientInfo
+from google.cloud import spanner_v1 as spanner
 
-from .cursor import Cursor
-from .exceptions import InterfaceError
-
-# from .version import google_client_info
-# from google.cloud.spanner_dbapi import Cursor
-# from google.cloud.spanner_dbapi import google_client_info
-# from google.cloud.spanner_dbapi import InterfaceError
+from google.cloud.spanner_dbapi.cursor import Cursor
+from google.cloud.spanner_dbapi.exceptions import InterfaceError
+from google.cloud.spanner_dbapi.version import DEFAULT_USER_AGENT
+from google.cloud.spanner_dbapi.version import PY_VERSION
 
 
 class Connection(object):
@@ -85,56 +83,54 @@ class Connection(object):
         self.close()
 
 
-# def connect(
-#     instance_id, database_id, project=None, credentials=None, user_agent=None
-# ):
-#     """Creates a connection to a Google Cloud Spanner database.
-#
-#     :type instance_id: str
-#     :param instance_id: ID of the instance to connect to.
-#
-#     :type database_id: str
-#     :param database_id: The name of the database to connect to.
-#
-#     :type project: str
-#     :param project: (Optional) The ID of the project which owns the
-#                     instances, tables and data. If not provided, will
-#                     attempt to determine from the environment.
-#
-#     :type credentials: :class:`~google.auth.credentials.Credentials`
-#     :param credentials: (Optional) The authorization credentials to attach to
-#                         requests. These credentials identify this application
-#                         to the service. If none are specified, the client will
-#                         attempt to ascertain the credentials from the
-#                         environment.
-#
-#     :type user_agent: str
-#     :param user_agent: (Optional) Prefix to the user agent header.
-#
-#     :rtype: :class:`google.cloud.spanner_dbapi.connection.Connection`
-#     :returns: Connection object associated with the given Google Cloud Spanner
-#               resource.
-#
-#     :raises: :class:`ValueError` in case of given instance/database
-#              doesn't exist.
-#     """
-#     client_info = google_client_info(user_agent)
-#
-#     client = spanner_v1.Client(
-#         project=project,
-#         credentials=credentials,
-#         # client_info=google_client_info(user_agent),
-#         client_info=client_info,
-#     )
-#
-#     instance = client.instance(instance_id)
-#     if not instance.exists():
-#         raise ValueError("instance '%s' does not exist." % instance_id)
-#
-#     database = instance.database(
-#         database_id, pool=spanner_v1.pool.BurstyPool()
-#     )
-#     if not database.exists():
-#         raise ValueError("database '%s' does not exist." % database_id)
-#
-#     return Connection(database)
+def connect(
+    instance_id, database_id, project=None, credentials=None, user_agent=None
+):
+    """Creates a connection to a Google Cloud Spanner database.
+
+    :type instance_id: str
+    :param instance_id: ID of the instance to connect to.
+
+    :type database_id: str
+    :param database_id: The name of the database to connect to.
+
+    :type project: str
+    :param project: (Optional) The ID of the project which owns the
+                    instances, tables and data. If not provided, will
+                    attempt to determine from the environment.
+
+    :type credentials: :class:`~google.auth.credentials.Credentials`
+    :param credentials: (Optional) The authorization credentials to attach to
+                        requests. These credentials identify this application
+                        to the service. If none are specified, the client will
+                        attempt to ascertain the credentials from the
+                        environment.
+
+    :type user_agent: str
+    :param user_agent: (Optional) Prefix to the user agent header.
+
+    :rtype: :class:`google.cloud.spanner_dbapi.connection.Connection`
+    :returns: Connection object associated with the given Google Cloud Spanner
+              resource.
+
+    :raises: :class:`ValueError` in case of given instance/database
+             doesn't exist.
+    """
+
+    client_info = ClientInfo(
+        user_agent=user_agent or DEFAULT_USER_AGENT, python_version=PY_VERSION,
+    )
+
+    client = spanner.Client(
+        project=project, credentials=credentials, client_info=client_info,
+    )
+
+    instance = client.instance(instance_id)
+    if not instance.exists():
+        raise ValueError("instance '%s' does not exist." % instance_id)
+
+    database = instance.database(database_id, pool=spanner.pool.BurstyPool())
+    if not database.exists():
+        raise ValueError("database '%s' does not exist." % database_id)
+
+    return Connection(database)
