+++ tests/spanner_dbapi/test_globals.py
@@ -6,11 +6,15 @@
 
 from unittest import TestCase
 
-from google.cloud.spanner_dbapi import apilevel, paramstyle, threadsafety
-
 
 class DBAPIGlobalsTests(TestCase):
     def test_apilevel(self):
+        from google.cloud.spanner_dbapi import (
+            apilevel,
+            paramstyle,
+            threadsafety,
+        )
+
         self.assertEqual(apilevel, "2.0", "We implement PEP-0249 version 2.0")
         self.assertEqual(paramstyle, "format", "Cloud Spanner uses @param")
         self.assertEqual(
